shell/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
asyncpg==0.29.0
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pydantic[email]==2.5.3
